move NatMethod to domain package
