+++ besu/src/main/java/org/hyperledger/besu/cli/DefaultCommandValues.java
@@ -17,7 +17,7 @@ package org.hyperledger.besu.cli;
 import org.hyperledger.besu.ethereum.core.Wei;
 import org.hyperledger.besu.ethereum.eth.sync.SyncMode;
 import org.hyperledger.besu.ethereum.p2p.config.RlpxConfiguration;
-import org.hyperledger.besu.nat.NatMethod;
+import org.hyperledger.besu.nat.core.domain.NatMethod;
 import org.hyperledger.besu.util.bytes.BytesValue;
 
 import java.io.File;
