+++ nat/src/main/java/org/hyperledger/besu/nat/NatService.java
@@ -15,6 +15,7 @@
 package org.hyperledger.besu.nat;
 
 import org.hyperledger.besu.nat.core.NatManager;
+import org.hyperledger.besu.nat.core.domain.NatMethod;
 import org.hyperledger.besu.nat.core.domain.NatPortMapping;
 import org.hyperledger.besu.nat.core.domain.NatServiceType;
 import org.hyperledger.besu.nat.core.domain.NetworkProtocol;
