+++ nat/src/main/java/org/hyperledger/besu/nat/core/NatManager.java
@@ -14,7 +14,7 @@
  */
 package org.hyperledger.besu.nat.core;
 
-import org.hyperledger.besu.nat.NatMethod;
+import org.hyperledger.besu.nat.core.domain.NatMethod;
 import org.hyperledger.besu.nat.core.domain.NatPortMapping;
 import org.hyperledger.besu.nat.core.domain.NatServiceType;
 import org.hyperledger.besu.nat.core.domain.NetworkProtocol;
