+++ nat/src/main/java/org/hyperledger/besu/nat/core/domain/NatServiceType.java
@@ -17,7 +17,7 @@ package org.hyperledger.besu.nat.core.domain;
 
 /**
  * This enum describes the types of services that could be impacted by the {@link
- * org.hyperledger.besu.nat.NatMethod} used by the Besu node.
+ * NatMethod} used by the Besu node.
  *
  * <ul>
  *   <li><b>JSON_RPC:</b> Ethereum JSON-RPC HTTP service.
